Setup seperate test and development enviornments
In server/config/config.js some default values are set.
The "test" script sets a value of "test" for NODE_ENV.
If NODE_ENV has no value, then it gets a default value of
"development". A production setting should give it the value of
"production", which would guide it past the if/else block that sets
a default port and mongoose uri.