Associated users with todos and updated tests
The user now has a _creator field, which is set to the _id of the
user who created that todo. Now the routes are "privatized" so that
only the user who created a todo can interact with it.

The tests are updated to reflect the changes in the code and also
new ones were added to ensure that only a todo's creator can interact
with it.