Updated config to use JSON file
The "secret" for the jwt tokens is now a process.env property.
The secret is in a .json file which is included in .gitignore, so
that the secret data is not publicly visible. This handles both
test and development cases. For production, set the enviornment
variable then restart the server.
  Ex. heroku config:set JWT_SECRET=SomeRandomString20#!)$Jn2ps24